refactor to more easily find components
