+++ spring-back-end/pjl-application/pjl-application-core/src/main/java/com/pjlsoftware/SpringBackEndApplication.java
@@ -1,16 +1,14 @@
-package com.pjlsoftware.main;
+package com.pjlsoftware;
 
 import jakarta.annotation.PostConstruct;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
-import org.springframework.context.annotation.ImportResource;
 import org.springframework.data.jpa.convert.threeten.Jsr310JpaConverters;
 
 import java.util.TimeZone;
 
 @SpringBootApplication
-@ImportResource({"classpath:applicationContext.xml"})
 @EntityScan(basePackageClasses = {
         SpringBackEndApplication.class,
         Jsr310JpaConverters.class
